#  mimalloc as global allocator