--- a/mimium-lang/Cargo.toml
+++ b/mimium-lang/Cargo.toml
@@ -7,6 +7,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 # [lib]
 
+[features]
+default = ["mimalloc"]
+mimalloc = ["dep:mimalloc"]
+
 
 [dependencies]
 
@@ -18,3 +22,4 @@ slotmap = "1.0.7"
 colog = "1.3.0"
 half = "2.4.1"
 itertools = "0.13.0"
+mimalloc = { version = "0.1.43", optional = true }