--- a/mimium-lang/benches/bench.rs
+++ b/mimium-lang/benches/bench.rs
@@ -45,33 +45,65 @@ fn dsp(){{
             )
         }
 
-        fn bench_multiosc(b: &mut Bencher, n: usize) {
-            let content = make_multiosc_src(n);
-            let compiler = compiler::Context::new([].into_iter(), None);
-            let program = compiler.emit_bytecode(&content).expect("ok");
+        fn bench_runtime(b: &mut Bencher, content: &str, times: usize) {
+            let compiler = compiler::Context::new([], None);
+            let program = compiler.emit_bytecode(content).expect("ok");
             let idx = program.get_fun_index(&"dsp".to_symbol()).expect("ok");
             let mut machine = Machine::new(program, [].into_iter(), [].into_iter());
             machine.execute_main();
-            b.iter(move || machine.execute_idx(idx));
+            b.iter(move || {
+                for _i in 0..times {
+                    let _ = machine.execute_idx(idx);
+                }
+            });
         }
         #[bench]
         fn bench_multiosc5(b: &mut Bencher) {
-            bench_multiosc(b, 5);
+            bench_runtime(b, &make_multiosc_src(5),1);
         }
         #[bench]
         fn bench_multiosc7(b: &mut Bencher) {
-            bench_multiosc(b, 7);
+            bench_runtime(b, &make_multiosc_src(7),1);
         }
         #[bench]
         fn bench_multiosc10(b: &mut Bencher) {
-            bench_multiosc(b, 10);
+            bench_runtime(b, &make_multiosc_src(9),1);
         }
         #[bench]
         fn bench_multiosc15(b: &mut Bencher) {
-            bench_multiosc(b, 15);
+            bench_runtime(b, &make_multiosc_src(15),1);
+        }
+        fn make_partialapp_src_from_template(c: &str) -> String {
+            format!(
+                "let pi = 3.14159265359
+let sr = 44100.0
+fn phasor(freq,phase){{
+  (self + freq/sr + phase)%1.0
+}}
+fn osc(freq,phase){{
+  sin(phasor(freq,phase)*pi*2.0)
+}}
+fn dsp(){{
+    {c}
+}}"
+            )
+        }
+        fn make_partialapp_src() -> String {
+            make_partialapp_src_from_template("440 |> osc(_,0.0)")
+        }
+        fn make_no_partialapp_src() -> String {
+            make_partialapp_src_from_template("osc(440,0.0)")
+        }
+        //test the performance degradation when open closure is made on `dsp` function with partial application.
+        #[bench]
+        fn bench_partialapp(b: &mut Bencher) {
+            bench_runtime(b, &make_partialapp_src(),10);
+        }
+        #[bench]
+        fn bench_partialapp_no(b: &mut Bencher) {
+            bench_runtime(b, &make_no_partialapp_src(),10);
         }
     }
-
     mod parse {
         use mimium_lang::compiler;
         use test::Bencher;