@tomoyanonymous 2024-11-14
I'm going to merge this though there was not much significant performance improvement.
Maybe we should prepare the test case that hit to the worst-case performance of the allocator somehow.